Alligator Data

Dataset from Nifong et al 2015 which contains 2 food sources, 181 individuals and 2 isotopes. This dataset includes multiple covariates as well as TDF means and sds.

alligator_data

Format

A list with the following elements

mixtures

A two column matrix containing delta 13C and delta 15N values respectively

ID

A character vector of unique ID values

tag_ID

A character vector of tag ID

lat

A numeric vector of latitude

long

A numeric vector of longitude

date

A character vector of date

year

A numeric vector of year

habitat

A character vector of habitat

sex

A character vector of sex

length

A numeric vector of length in cm

s_class

A character vector of size class

sex_sclass

A character vector for sex times size class

source_names

A character vector of source names

source_means

A data frame of source means for same tracers as in Mixtures

source_sds

A data frame of standard deviations of sources for same tracers as in Mixtures

n_sources

A numeric vector of number of sources

TEF_means

A data frame of means for TEFs for same tracers as in Mixtures

TEF_sds

A data frame of sds for TEFs for same tracers as in Mixtures

Source

<doi:10.1111/1365-2656.12306>