Plot different features of an object created from cosimmr_ffvb.

This function allows for 4 different types of plots of the simmr output created from cosimmr_ffvb. The types are: plot of beta values

# S3 method for cosimmr_output
plot(
  x,
  type = c("isospace", "beta_histogram", "beta_boxplot", "p_ind"),
  ind = 1,
  covariates = 1,
  binwidth = 0.1,
  alpha = 0.5,
  title = NULL,
  n_output = 3600,
  ...
)

Arguments

x

An object of class cosimmr_output created via cosimmr_ffvb.

type

The type of plot required. Can be one or more of 'isospace', 'betahistogram'

ind

The individual number you wish to plot

covariates

The covariate you wish to plot (for beta plots)

binwidth

The width of the bins for the histogram. Defaults to 0.05

alpha

The degree of transparency of the plots. Not relevant for matrix plots

title

The title of the plot.

n_output

The number of theta samples you wish to plot with. Defaults to 3600

...

Currently not used

Details

The matrix plot should form a necessary part of any SIMM analysis since it allows the user to judge which sources are identifiable by the model. Further detail about these plots is provided in the vignette.

See also

Author

Emma Govan <emmagovan@gmail.com>, Andrew Parnell

Examples

if (FALSE) {
# A simple example with 10 observations, 2 tracers and 4 sources

# The data
data(geese_data_day1)

# Load into simmr
simmr_1 <- with(
  geese_data_day1,
  cosimmr_load(
    formula = mixtures ~ 1,
    source_names = source_names,
    source_means = source_means,
    source_sds = source_sds,
    correction_means = correction_means,
    correction_sds = correction_sds,
    concentration_means = concentration_means
  )
)
# Plot
plot(simmr_1)

# FFVB run
simmr_1_out <- cosimmr_ffvb(simmr_1)

plot(simmr_1_out, type = c("isospace", "beta_hist"))
}